pyproject.toml
[tool.poetry]
name = "kodinerds-iptv"
version = "0.1.0"
description = "Free TV and radio streams for Kodi."
authors = ["Jannik Schäfer <jannik.schaefer@inbox-online.de>"]
readme = "README.md"
packages = [{ include = "kodinerds_iptv" }]

[tool.poetry.dependencies]
# TODO: Support more Python versions if possible.
python = "^3.11"
PyYAML = "^6.0"
httpx = "^0.26.0"
typer = { extras = ["all"], version = "^0.9.0" }
tabulate = "^0.9.0"
aiometer = "^0.5.0"
loguru = "^0.7.2"

[tool.poetry.scripts]
kn-iptv = "kodinerds_iptv.main:app"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
types-pyyaml = "6.0.12.12"
types-tabulate = "0.9.0.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.coverage.report]
exclude_lines = ["if (typing\\.)?TYPE_CHECKING:"]

[tool.mypy]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
select = ["ALL"]
target-version = "py311"
ignore = ["ANN101", "ANN102", "COM812", "TD002", "TD003"]
unfixable = ["T201", "ERA001", "F841"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["D104", "F401"]
"tests/*.py" = ["S101"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.vulture]
min_confidence = 100
paths = ["kodinerds_iptv"]
sort_by_size = true
